modify code.

+++ rs/src/main/java/com/higgs/trust/rs/core/bo/ContractQueryRequestV2.java
@@ -13,7 +13,7 @@ public class ContractQueryRequestV2 {
     /**
      * Height of block.
      */
-    private int blockHeight;
+    private long blockHeight;
     /**
      * Contract address, hex string of 40 characters, e.g. 00a615668486da40f31fd050854fb137b317e056.
      */
